+++ lib/src/errors.dart
@@ -14,8 +14,6 @@
  * # limitations under the License.
  */
 
-// ignore_for_file: avoid_as
-
 library gltf.error;
 
 import 'package:meta/meta.dart';
@@ -619,6 +617,10 @@ class LinkError extends IssueType {
               'for this attribute semantic. '
               'Must be one of ${(args[1] as Iterable).map(_q)}.');
 
+  static final LinkError meshPrimitiveAttributesAccessorUnsignedInt =
+      LinkError._('MESH_PRIMITIVE_ATTRIBUTES_ACCESSOR_UNSIGNED_INT',
+          (args) => 'Mesh attributes cannot use UNSIGNED_INT component type.');
+
   static final LinkError meshPrimitivePositionAccessorWithoutBounds =
       LinkError._(
           'MESH_PRIMITIVE_POSITION_ACCESSOR_WITHOUT_BOUNDS',
@@ -725,10 +727,14 @@ class LinkError extends IssueType {
       'UNSUPPORTED_EXTENSION',
       (args) => 'Cannot validate an extension as it is not supported '
           'by the validator: ${_q(args[0])}.',
-      Severity.Warning);
+      Severity.Information);
 
   static final LinkError unusedObject = LinkError._('UNUSED_OBJECT',
       (args) => 'This object may be unused.', Severity.Information);
+
+  static final LinkError khrMaterialsVariantsNonUniqueVariant = LinkError._(
+      'KHR_MATERIALS_VARIANTS_NON_UNIQUE_VARIANT',
+      (args) => 'This variant is used more than once for this mesh primitive.');
 
   LinkError._(String type, ErrorFunction message,
       [Severity severity = Severity.Error])