+++ test/ext/KHR_materials_transmission/khr_materials_transmission_test.dart
@@ -9,8 +9,6 @@ import 'package:gltf/gltf.dart';
 
 import '../../utils.dart';
 
-//ignore_for_file: avoid_as
-
 Future main() async {
   await compareReports('test/ext/KHR_materials_transmission');